# vityaz-main/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/orders": {
            "get": {
                "operationId": "OrdersController_getOrders",
                "summary": "Получение всех заказов",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Получение всех заказов текущего пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderDto"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Заказы"
                ]
            },
            "post": {
                "operationId": "OrdersController_createOrder",
                "summary": "Создание заказа",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "Перечисление элементов из корзины, которые будут в заказе. Поле может быть пустым - тогда в заказ войдут все предметы из корзины",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartItemsIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заказ был успешно создан"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Заказы"
                ]
            }
        },
        "/api/orders/{orderId}": {
            "get": {
                "operationId": "OrdersController_getOrder",
                "summary": "Получение конкретного заказа",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderId",
                        "required": true,
                        "in": "path",
                        "description": "Идентификатор заказа",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Получение указанного заказа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderDto"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "NotFoundException": {
                                        "description": "Заказ с указанным идентификатором не был найден",
                                        "value": {
                                            "statusCode": 404,
                                            "message": "Order not found",
                                            "error": "Not Found"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Заказ с указанным идентификатором не был найден",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order not found"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Заказы"
                ]
            },
            "post": {
                "operationId": "OrdersController_changeOrderStatus",
                "summary": "Изменение статуса заказа",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderId",
                        "required": true,
                        "in": "path",
                        "description": "Идентификатор заказа",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "Новый статус заказа",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderStatusDto"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Статус заказа успешно изменяется"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "UnauthorizedException": {
                                        "description": "Пользователь поставил некорректный токен доступа",
                                        "value": {
                                            "statusCode": 401,
                                            "message": "Invalid access token",
                                            "error": "Unauthorized"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь поставил некорректный токен доступа",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid access token"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "ForbiddenException": {
                                        "description": "Пользователь не имеет прав выполнять данную операцию",
                                        "value": {
                                            "statusCode": 403,
                                            "message": "You dont have enough rights to perform this operation",
                                            "error": "Forbidden"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Пользователь не имеет прав выполнять данную операцию",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 403
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You dont have enough rights to perform this operation"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "NotFoundException": {
                                        "description": "Заказ с указанным идентификатором не был найден",
                                        "value": {
                                            "statusCode": 404,
                                            "message": "Order not found",
                                            "error": "Not Found"
                                        }
                                    }
                                },
                                "schema": {
                                    "type": "object",
                                    "description": "Заказ с указанным идентификатором не был найден",
                                    "properties": {
                                        "statusCode": {
                                            "type": "number",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order not found"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "required": [
                                        "statusCode",
                                        "message"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    }
                },
                "tags": [
                    "Заказы",
                    "Для администратора"
                ]
            }
        },
...
    },
    "info": {
        "title": "Магазин «Витязь»",
        "description": "",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "ProductDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "origin": {
                        "type": "string"
                    },
                    "picture": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "name",
                    "price",
                    "brand",
                    "origin",
                    "picture"
                ]
            },
...
            "OrderItemDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Идентификатор элемента",
                        "readOnly": true
                    },
                    "product": {
                        "description": "Товар",
                        "readOnly": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/ProductDto"
                            }
                        ]
                    },
                    "quantity": {
                        "type": "number",
                        "description": "Количество товара",
                        "readOnly": true
                    },
                    "price": {
                        "type": "number",
                        "description": "Цена позиции",
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "quantity",
                    "price"
                ]
            },
            "OrderDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Уникальный идентификатор заказа пользователя",
                        "nullable": false,
                        "readOnly": true
                    },
                    "items": {
                        "description": "Элементы заказа",
                        "nullable": false,
                        "readOnly": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemDto"
                        }
                    },
                    "totalPrice": {
                        "type": "number",
                        "description": "Общая стоимость товаров в заказе",
                        "nullable": false,
                        "readOnly": true
                    }
                },
                "required": [
                    "id",
                    "items",
                    "totalPrice"
                ]
            },
            "OrderStatusDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "description": "Новый статус заказа",
                        "writeOnly": true,
                        "nullable": false,
                        "examples": [
                            [
                                "created",
                                "processing",
                                "sent",
                                "delivered",
                                "rejected"
                            ]
                        ]
                    }
                },
                "required": [
                    "status"
                ]
            },
            "CartItemsIdsDto": {
                "type": "object",
                "properties": {
                    "ids": {
                        "description": "Перечисление идентификаторов элементов, которые попадут в заказ",
                        "type": "array",
                        "items": {
                            "type": "array"
                        }
                    }
                }
            },
...
        }
    }
}